% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_average.R
\name{angle_average}
\alias{angle_average}
\title{Circular average of a set of angle degrees}
\usage{
angle_average(data)
}
\arguments{
\item{data}{a numeric vector of angle degrees.}
}
\value{
a number, the mean degree from the input \code{data}.
}
\description{
Calculating the circular average from a set of angle degrees.
}
\examples{
# calculate the average angle of the three: 350, 5, 20
deg.ave = angle_average(data = c(350, 5, 20))
deg.ave
}
