% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emd_simple.R
\name{emd_simple}
\alias{emd_simple}
\title{The Earth mover's distance between two sequences with the same length.}
\usage{
emd_simple(data)
}
\arguments{
\item{data}{a data.frame of observations by different stations, each
column represents a station, and the rows represent standard recording datetimes.}
}
\value{
a matrix, the value at row i and column j is the Earth
mover's distance between station i and j in `data`.
}
\description{
Calculating the Earth mover's distance (Wasserstein L-1) between
the observations of any two stations in the input data, which have the same
reporting datetimes.
}
\examples{
wow_emd = emd_simple(wow_ws)
}
