% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqm_bc_5foldcv_simple.R
\name{eqm_bc_5foldcv_simple}
\alias{eqm_bc_5foldcv_simple}
\title{Calibrating the bias of observed data by empirical quantile
mapping method with 5-fold cross-validation}
\usage{
eqm_bc_5foldcv_simple(data.obs, data.datetime, true.quantiles)
}
\arguments{
\item{data.obs}{a numeric vector of observation data needs to be corrected.}

\item{data.datetime}{a sequence of timestamps in \code{data.obs}.}

\item{true.quantiles}{a list of six numeric vectors of the "truth" quantiles to build the
bias correction model, the corresponding cumulative probability stamps must be evenly spreaded
between 0 and 1. If not satisfied,
a pseudo "truth" observation numeric vector satisfying the quantiles should be provided.
We suggest using quantiles with cumulative probability from 0.01 to 1 with interval length 0.01
(100 quantiles in total).
We sort these quantiles to avoid crossovers.}
}
\value{
a list of two parts: one is the output dataframe \code{bc_output} that includes both
input and bias-corrected observation data; the other one is a large list \code{bc.eqm_model}
contains all the empirical quantile mapping models in each turn of cross-validation
during different season/period.
}
\description{
Performing empirical quantile distribution mapping to correct bias in the observation data,
applying the 5-fold cross-validation;
the "truth" quantiles are given for a reference.
}
