% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_zeros.R
\name{estimate_zeros}
\alias{estimate_zeros}
\title{Estimating zeros}
\usage{
estimate_zeros(
  test_windspeed.afterSpatialQC.nonzero,
  test.station.windspeed,
  official.station.windspeed,
  test.nbhd_spatial.info
)
}
\arguments{
\item{test_windspeed.afterSpatialQC.nonzero}{a data.frame that includes observation data after spatial
consistency check, with no zeros.}

\item{test.station.windspeed}{a data.frame that includes observation data of test stations before
spatial consistency check.}

\item{official.station.windspeed}{a data.frame that includes observation data of official stations.}

\item{test.nbhd_spatial.info}{output of function \code{spatial_neighbors}.}
}
\value{
a new data.frame that includes wind speed observations after estimating zeros.
}
\description{
Provide estimates of zero observations cause by systematic under-estimation, using
inverse Earth mover's distance weighting method for interpolations.
}
