% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors_selection.R
\name{neighbors_selection}
\alias{neighbors_selection}
\title{Select suitable neighboring stations to do spatial quality control}
\usage{
neighbors_selection(data, ref.data, geo.distance = FALSE)
}
\arguments{
\item{data}{a data.frame of observations by different stations, each
column represents a station, and the rows represent standard recording datetimes.}

\item{ref.data}{a data.frame that has the same settings with `data`; the stations
in `ref.data` are official stations used for reference.}

\item{geo.distance}{`TURE`/`FALSE`; The default is `FALSE`, which means the neighbors
selection function do not consider geographical distances between stations; If setting
`TRUE`, the longitude and latitude information of stations in `data` and `ref.data`
should be provided to calculate the geographical distances.}
}
\value{
a list and a vector; In the list, each element contains the six choosen
neighboring stations from either `data` or `ref.data` for a candidate station in `data`;
The vector indicated the best order of stations in `data` to do spatial quality control,
a station that has more official stations in the neighboring list
for reference would be checked first in the spatial quality control.
}
\description{
Choosing neighboring stations mainly based on the similarities
(both Pearson correlation and Earth mover's distance0) between wind speed
distributions of the input data; Also get the optimal order of candidate
stations to do spatial quality control.
}
\examples{
wow_nbhd = neighbors_selection(wow_ws, knmi_ws)[[1]]
wow_order = neighbors_selection(wow_ws, knmi_ws)[[2]]
}
