% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCwind-package.R
\docType{package}
\name{QCwind-package}
\alias{QCwind}
\alias{QCwind-package}
\title{QCwind: Quality control procedures for citizen science wind observations}
\description{
This package contains the functions used to do quality control and bias correction for wind data.
}
\keyword{internal}
