% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_check.R
\name{range_check}
\alias{range_check}
\title{Range check function}
\usage{
range_check(data, column, upper.bound, lower.bound = 0, fail.flag)
}
\arguments{
\item{data}{a data.frame that includes observation data with information.}

\item{column}{a character, the selected column name in the data.frame that
represents observation data to be checked.}

\item{upper.bound}{a real number that controls the maximum allowed value of observations.}

\item{lower.bound}{a real number that controls the minimum required value of observations (default value = 0).}

\item{fail.flag}{a character/string that represents the name of flag where an observation fails the range check.}
}
\value{
a new tbl_df that extends two new columns on the input data,
the first added column `flag` represents the flags (`P` means pass, `fail.flag` means fail,
`missing` means missing values of observations), the second added column `new_data` represents
the observation data after removing failed observations.
}
\description{
Checking the plausible range of a sequence of observations;
The observations should be no less than the `lower.bound` and no more than the `upper.bound`;
Generating a new sequence of flags, where a `P` means that observation passes the range check,
and a pre-defined flag name `fail.flag` means that observation fails the range check;
Also outputting a new sequence of observations after removing failed observations.
}
\examples{
wow_test_range_check = range_check(wow_test, 'windspeed_metrepersecond',
upper.bound = 25, fail.flag = 'RS')
attributes(wow_test_range_check)
}
