% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_knmi_gz_data.R
\name{read_knmi_gz_data}
\alias{read_knmi_gz_data}
\title{Transform/Split KNMI gziped wind data file}
\usage{
read_knmi_gz_data(raw)
}
\arguments{
\item{raw}{the imported gzipped file with only one column contains all varaible values}
}
\value{
a splitted dataframe/matrix with multiple columns, each one stands for one variable
}
\description{
Transform/Split KNMI gzipped wind observations date file to a dataframe/matrix
}
