% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_check.R
\name{spatial_check}
\alias{spatial_check}
\title{Spatial consistency check using inverse Earth mover's distance weighting method}
\usage{
spatial_check(
  test.station.windspeed.nonzero,
  official.station.windspeed,
  test.nbhd_spatial.info,
  order_spatial.test_station
)
}
\arguments{
\item{test.station.windspeed.nonzero}{a data.frame that includes observation data of all test
stations, with no zeros.}

\item{official.station.windspeed}{a data.frame that includes observation data of official stations.}

\item{test.nbhd_spatial.info}{output of function \code{spatial_neighbors}.}

\item{order_spatial.test_station}{output of function \code{spatial_order}.}
}
\value{
a large list of three data.frames:
new_windspeed: observations after filtering spatially inconsistent ones;
flags: flag of each observation;
iemdw_estimates: inverse Earth mover's distance weighting estimates of each observation.
}
\description{
Flag and reject inconsistent observations in space.
}
