% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_neighbors.R
\name{spatial_neighbors}
\alias{spatial_neighbors}
\title{Selecting neighbors for each test station}
\usage{
spatial_neighbors(
  test.station.windspeed,
  official.station.windspeed,
  test.station.info,
  official.station.info,
  emd_radius = 1
)
}
\arguments{
\item{test.station.windspeed}{a data.frame that includes observation data of all test stations.}

\item{official.station.windspeed}{a data.frame that includes observation data of official stations.}

\item{test.station.info}{a tbl_df that includes 'station_id', 'longitude', 'latitude' of test stations,
corresponding to their column number in \code{test.station.windspeed}.}

\item{official.station.info}{a tbl_df that includes 'location', 'longitude', 'latitude' of official
stations, corresponding to their column number in \code{official.station.windspeed}.}

\item{emd_radius}{the maximum radius of Earth mover's distance, setting 1 as default.}
}
\value{
a large list, each element is a data.frame that contains information of selected neighbors.
}
\description{
Selecting neighbors for each test station according to geographical distance,
correlation, Earth mover's distance.
}
