% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_order.R
\name{spatial_order}
\alias{spatial_order}
\title{Implementing order of test stations}
\usage{
spatial_order(test_station.scores, test.nbhd_spatial.info)
}
\arguments{
\item{test_station.scores}{a data.frame that includes the quality score of test stations.}

\item{test.nbhd_spatial.info}{output of function \code{spatial_neighbors}.}
}
\value{
a numeric sequence that shows the order of test stations in their column numbers.
}
\description{
Find an optimal order of test stations when performing the spatial consistency check.
}
