% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_qc.R
\name{spatial_qc}
\alias{spatial_qc}
\title{Spatial quality control of input observations}
\usage{
spatial_qc(data, ref.data, data.nbhd, data.order, fail.flag, method = "idw")
}
\arguments{
\item{data}{a data.frame of observations by different stations, each
column represents a station, and the rows represent standard recording datetimes.}

\item{ref.data}{a data.frame that has the same settings with `data`; the stations
in `ref.data` are official stations used for reference.}

\item{data.nbhd}{a list, the selected neighboring stations of each station in `data`
(obtained by the function `neighbors_selection`).}

\item{data.order}{a vector that indicating the optimal order of stations in `data`
to do spatial quality control (obtained by the function `neighbors_selection`).}

\item{fail.flag}{a character/string that represents the name of flag where an
observation fails the spatial quality control.}

\item{method}{a character indicating the spatial quality control method to be used;
`idw`: the inverse distance weighting method, this is also the default;
`srt`: the spatial weighted regression test (! needed much longer processing time).}
}
\value{
a list of four data.frames that have the same size with `data`:
`estimation`: the estimation of each target observation generated by neighboring reference
observations;
`nbhd_sd`: the standard deviations of neighboring reference observations related to
each target observation;
`flag`: the flagging information that indicates whether a target observation passes
spatial quality control of not, `P` is a pass, `fail.flag` is a fail, `isolated` represent
isolated observations;
`new_data`: the new data of the input `data` that removes the failed observations.
}
\description{
Implementing the spatial quality control, and receive the information
of which observations faild the quality control.
}
\examples{
wow_spatial = spatial_qc(wow_ws_test, knmi_ws_test, wow_nbhd, wow_order, fail.flag = 'F')
}
