% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_persist_check.R
\name{temporal_persist_check}
\alias{temporal_persist_check}
\title{Temporal persistence check function}
\usage{
temporal_persist_check(
  data,
  data.column,
  datetime.column,
  persist.duration,
  min.variation,
  realtime = FALSE
)
}
\arguments{
\item{data}{a data.frame that includes observation data with information.}

\item{data.column}{a character, the selected column name in the data.frame that
represents observation data to be checked.}

\item{datetime.column}{a character, the selected column name in the data.frame that
represents the reporting dates and times of corresponding observations.}

\item{persist.duration}{a positive number, the interval time for persistence test,
the unit is 'second'.}

\item{min.variation}{a positive number, the minimum required variability of observations.}

\item{realtime}{a logical. \code{TRUE} means processing in realtime,
the attributes that includes pass percent would be removed in realtime case.}
}
\value{
a new tbl_df that extends two new columns on the input data,
the first added column \code{flag_persist} represents the flags (\code{P} means pass, \code{fail.persist} means fail,
\code{isolated} means missing previous observations), the second added column \code{new_data_persist} represents
the observation data after removing failed observations.
}
\description{
Check on a minimum required variability of an instantaneous value;
Observations should change more than a minimum required number \code{min.variation},
during a time interval of length \code{persist.duration};
Generating a new sequence of flags, and outputting a new sequence of observations after removing failed observations.
}
\examples{
library(tidyverse)
datetime = as.POSIXlt(seq(0,60000,600), origin = "2017-02-03 08:00:00")
test_data = tibble(datetime = datetime,
              windspeed = c(rep(10.2,5),rep(11.3,15),seq(10,15,0.1),
                            rep(9.3,20),c(NA,NA,NA,2,NA,5,7,NA,2,4)))
test_persist_check = temporal_persist_check(test_data, data.column = 'windspeed',
                                            datetime.column = 'datetime',
                                            persist.duration = 3600,
                                            min.variation = 0.1)
attributes(test_persist_check)
test_persist_check
}
