% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_step_check.improve.R
\name{temporal_step_check.improve}
\alias{temporal_step_check.improve}
\title{Temporal step check function (new version, can be implemented in realtime)}
\usage{
temporal_step_check.improve(
  data,
  data.column,
  datetime.column,
  step.duration,
  max.variation
)
}
\arguments{
\item{data}{a data.frame that includes observation data with information.}

\item{data.column}{a character, the selected column name in the data.frame that
represents observation data to be checked.}

\item{datetime.column}{a character, the selected column name in the data.frame that
represents the reporting dates and times of corresponding observations.}

\item{step.duration}{a positive number, the interval time for step test, the unit is 'second'.}

\item{max.variation}{a positive number, the maximum allowed variability of observations.}
}
\value{
a new tbl_df that extends two new columns on the input data,
the first added column \code{flag_step} represents the flags (\code{P} means pass, \code{fail.step} means fail,
\code{isolated} means missing previous observations), the second added column \code{new_data_step} represents
the observation data after removing failed observations in the step check.
}
\description{
Same as the description in \code{temporal_step_check}.
}
\examples{
library(tidyverse)
datetime = as.POSIXct(seq(0,60000,600), origin = "2017-02-03 08:00:00")
test_data = tibble(datetime = datetime, windspeed = c(rep(10.2,5),rep(NA,5),rep(9.3,35),rep(20,5),
                                                rep(3.5,40),c(1,3,NA,2,NA,5,6,7,NA,2,4)))
test_step_check = temporal_step_check.improve(test_data, data.column = 'windspeed',
                                      datetime.column = 'datetime',
                                      step.duration = 660,
                                      max.variation = 5)
attributes(test_step_check)
test_step_check
}
