% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_step_check.new.R
\name{temporal_step_check.new}
\alias{temporal_step_check.new}
\title{Temporal step check function (fast version, for 10-min observations time series only)}
\usage{
temporal_step_check.new(
  data,
  data.column,
  datetime.column,
  step.duration,
  max.variation
)
}
\arguments{
\item{data}{a data.frame that includes observation data with information.}

\item{data.column}{a character, the selected column name in the data.frame that
represents observation data to be checked.}

\item{datetime.column}{a character, the selected column name in the data.frame that
represents the reporting dates and times of corresponding observations.}

\item{step.duration}{a positive number, the interval time for step test, the unit is 'second'.}

\item{max.variation}{a positive number, the maximum allowed variability of observations.}
}
\value{
a new tbl_df that extends two new columns on the input data,
the first added column \code{flag_step} represents the flags (\code{P} means pass, \code{fail.step} means fail,
\code{isolated} means missing previous observations), the second added column \code{new_data_step} represents
the observation data after removing failed observations in the step check.
}
\description{
Same as the description in \code{temporal_step_check}.
}
