% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_step_check.R
\name{temporal_step_check}
\alias{temporal_step_check}
\title{Temporal step check function}
\usage{
temporal_step_check(
  data,
  data.column,
  datetime.column,
  step.duration,
  max.variation,
  fail.flag
)
}
\arguments{
\item{data}{a data.frame that includes observation data with information.}

\item{data.column}{a character, the selected column name in the data.frame that
represents observation data to be checked.}

\item{datetime.column}{a character, the selected column name in the data.frame that
represents the reporting dates and times of corresponding observations.}

\item{step.duration}{a positive number, the interval time for step test, the unit is 'second'.}

\item{max.variation}{a positive number, the maximum allowed variability of observations.}

\item{fail.flag}{a character/string that represents the name of flag where an
observation fails the step check.}
}
\value{
a new tbl_df that extends two new columns on the input data,
the first added column `flag` represents the flags (`P` means pass, `fail.flag` means fail,
`isolated` means missing previous observations), the second added column `new_data` represents
the observation data after removing failed observations.
}
\description{
Check on a maximum allowed variability of an instantaneous value;
Observations should change less than a maximum allowed number `max.variation`,
during a time interval of length `step.duration`;
Generating a new sequence of flags, where a `P` means that observation passes the step check,
and a pre-defined flag name `fail.flag` means that observation fails the step check;
Also outputting a new sequence of observations after removing failed observations.
}
\examples{
wow_test_step_check = temporal_step_check(wow_test_each[[1]],
data.column = 'windspeed_metrepersecond', datetime.column = 'datetime',
step.duration = 720, max.variation = 5, fail.flag = 'TS1')
attributes(wow_test_step_check)
}
