% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_data.R
\name{uniform_data}
\alias{uniform_data}
\title{Converting observation data to uniform time series}
\usage{
uniform_data(
  data,
  data.column,
  datetime.column,
  timeseq,
  freq = 600,
  method = "average"
)
}
\arguments{
\item{data}{a data.frame that includes observation data with datetime information.}

\item{data.column}{a character, the selected column name in the data.frame that
represents observation data to be checked.}

\item{datetime.column}{a character, the selected column name in the data.frame that
represents dates and times of observations to be checked.}

\item{timeseq}{a date and time vector, indicating the standard reporting timestamps.
`timeseq` must have the same length with `data`.}

\item{freq}{a number, the standard reporting interval time, the unit is "second".
We typically use 10 (min) for forecast use, and 60 (min) for analysis use.}

\item{method}{the matching method, choosing from "average", "maximum", "nearest", or "angle.average".
"average": the average value of observations in each time interval with length `freq`.
"maximum": the maximum value of observations in each time interval with length `freq`.
"nearest": the nearest observation to `timeseq` in each time interval with length `freq`.
"angle.average": the circular average value of angle observations
in each time interval with length `freq`.}
}
\value{
a new matched time series, with standard reporting time every `interval` minutes.
}
\description{
Citizen weather observations do not have uniform standards, this function converts the time series of
observations by various citizen weather stations into the same unified form.
}
