% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windrose_plot.R
\name{windrose_plot}
\alias{windrose_plot}
\title{Plotting a windrose}
\usage{
windrose_plot(data, speed, direction)
}
\arguments{
\item{data}{a dataframe that contains the wind speed and direction data}

\item{speed}{the wind speed column from data}

\item{direction}{the wind direction column from data}
}
\value{
a windrose plot
}
\description{
Plot a windrose of the given wind speed and direction data
}
