% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correction.R
\name{bias_correction}
\alias{bias_correction}
\title{Function to do bias correction}
\usage{
bias_correction(obs.data, true.data, method = "empirical_quantile")
}
\arguments{
\item{obs.data}{a numeric vector of observed data.}

\item{true.data}{a numeric vector of true data.}

\item{method}{a character indicating the choosen bias correction method;
`linear_scaling`: the linear scaling method, fitting the relationship between observed and
true data by a linear model;
`empirical_quantile`: the empirical quantile mapping method, mapping the quantiles between
the empirical cumulative density functions of observed and true data.}
}
\value{
a function that can correct new observed data.
}
\description{
Using different methods to calibrate the systematic bias between observed
and true data.
}
\examples{
wow_test_bc = bias_correction(bc_wow_test, bc_knmi_test, 'empirical_quantile')
}
