% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correction.R
\name{bias_correction}
\alias{bias_correction}
\title{Calibrating the bias of observed data}
\usage{
bias_correction(train.obs, train.true, test.obs, method = "empirical_quantile")
}
\arguments{
\item{train.obs}{a numeric vector of training observed data to build the bias correction model.}

\item{train.true}{a numeric vector of training true data to build the bias correction model.}

\item{test.obs}{a numeric vector of test observed data to be calibrated.}

\item{method}{a character indicating the choosen bias correction method;
`linear_scaling`: the linear scaling method, fitting the relationship between observed and
true data by a linear model;
`empirical_quantile`: the empirical quantile mapping method, mapping the quantiles between
the empirical cumulative density functions of observed and true data.}
}
\value{
a numeric vector that contains the corrected `test.obs` data.
}
\description{
Using different methods to calibrate the systematic bias between observed
and true data.
}
\examples{
wow_test_bc = bias_correction(bc_wow_test, bc_knmi_test, bc_wow_test)
}
