% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_persist_check.R
\name{temporal_persist_check}
\alias{temporal_persist_check}
\title{Temporal persistence check function}
\usage{
temporal_persist_check(
  data,
  data.column,
  datetime.column,
  persist.duration,
  min.variation,
  fail.flag
)
}
\arguments{
\item{data}{a data.frame that includes observation data with information.}

\item{data.column}{a character, the selected column name in the data.frame that
represents observation data to be checked.}

\item{datetime.column}{a character, the selected column name in the data.frame that
represents the reporting dates and times of corresponding observations.}

\item{persist.duration}{a positive number, the interval time for persistence test,
the unit is 'second'.}

\item{min.variation}{a positive number, the minimum required variability of observations.}

\item{fail.flag}{a character/string that represents the name of flag where an
observation fails the persistence check.}
}
\value{
a new tbl_df that extends two new columns on the input data,
the first added column `flag` represents the flags (`P` means pass, `fail.flag` means fail,
`isolated` means missing previous observations), the second added column `new_data` represents
the observation data after removing failed observations.
}
\description{
Check on a minimum required variability of an instantaneous value;
Observations should change more than a minimum required number `min.variation`,
during a time interval of length `persist.duration`;
Generating a new sequence of flags, where a `P` means that observation passes the persistence check,
and a pre-defined flag name `fail.flag` means that observation fails the persistence check;
Also outputting a new sequence of observations after removing failed observations.
}
\examples{
wow_test_persist_check = temporal_persist_check(wow_test_each[[1]],
data.column = 'windspeed_metrepersecond', datetime.column = 'datetime',
persist.duration = 3600, min.variation = 0.1, fail.flag = 'TS2')
attributes(wow_test_persist_check)
}
